Move sha3 over to optionalDependencies
`sha3@1.2.2` with it's C implementation is about 45x faster than `sha3@2.0.0`'s pure JS implementation, so I think it makes sense to keep 1.2.2 around (see https://github.com/phusion/node-sha3/pull/44#issuecomment-437689120 for benchmarks).

Also, `browserify-sha3` uses `js-sha3`, which is currently the fastest pure-js implementation, so it makes sense to keep it for the browser-fallback.

Additionally, many users installing this library have issues doing so due to sha3@1.2.2's dependency on node-gyp, python, and other missing build tools.

To work around this issue `sha3@1.2.2` can be installed as an `optionalDependency`, which will permit the install of `sha3` to fail, but the install of `keccakjs` to still succeed (just like is done with the `scrypt` dependency in `scrypt.js`).